feat(dump-agent-go): subcommand dispatcher (run, service, install, uninstall, version)
