+++ apps/dump_agent_go/cmd/dumpagent/cmd_version.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runVersion() int {
+	fmt.Printf("dumpagent %s\n", Version)
+	return 0
+}
